校验和修改

+++ X-fim-api/src/main/java/com/feige/api/codec/ICheckSum.java
@@ -12,8 +12,8 @@ public interface ICheckSum extends Spi {
      * @return Equal or not
      * @throws CheckSumException
      */
-    default void check(byte[] body, byte expectedCheckSum) throws CheckSumException {
-        byte checksumResult = getCheckSum(body);
+    default void check(byte[] body, short expectedCheckSum) throws CheckSumException {
+        short checksumResult = getCheckSum(body);
         if (checksumResult != expectedCheckSum) {
             throw new CheckSumException(String.format(
                     "stream corrupted: mismatching checksum: %d (expected: %d)",
@@ -26,5 +26,5 @@ public interface ICheckSum extends Spi {
      * @param body body
      * @return check sum
      */
-    byte getCheckSum(byte[] body);
+    short getCheckSum(byte[] body);
 }
